# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
    AdopterDto:
      type: object
      properties:
        repository:
          type: string
        catalogId:
          type: string
        repositoryStatus:
          type: string
        archetype:
          type: string
        depth:
          type: integer
          format: int32
        via:
          type: array
          items:
            type: string
        state:
          type: string
        adoptedVersion:
          type: string
        adoptedAt:
          $ref: "#/components/schemas/Instant"
    AdoptionJourneyDto:
      type: object
      properties:
        repository:
          type: string
        catalogId:
          type: string
        version:
          type: string
        packages:
          type: array
          items:
            $ref: "#/components/schemas/PackageDto"
        adopters:
          type: array
          items:
            $ref: "#/components/schemas/AdopterDto"
...
    ArtifactDto:
      type: object
      properties:
        ecosystem:
          type: string
        name:
          type: string
        repository:
          type: string
        latest:
          type: string
        version:
          type: string
        occurredAt:
          $ref: "#/components/schemas/Instant"
        sbomStatus:
          type: string
        dependentCount:
          type: integer
          format: int32
        behindCount:
          type: integer
          format: int32
...
    BumpDto:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        repository:
          type: string
        group:
          type: string
        mode:
          type: string
        branch:
          type: string
        environment:
          type: string
        trigger:
          type: string
        status:
          type: string
        ciEventId:
          type: string
        ciRunId:
          type: string
        ciRunIds:
          type: array
          items:
            type: string
        configPath:
          type: string
        ciRunStatus:
          type: string
        startedAt:
          $ref: "#/components/schemas/Instant"
        finishedAt:
          $ref: "#/components/schemas/Instant"
        message:
          type: string
        resultSha:
          type: string
        releaseRequestId:
          type: string
        releaseState:
          type: string
        releaseDetail:
          type: string
        releaseStateAt:
          $ref: "#/components/schemas/Instant"
        baseRef:
          type: string
        replaceHead:
          type: string
        changes:
          type: array
          items:
            $ref: "#/components/schemas/Change"
        failure:
          $ref: "#/components/schemas/FailureDto"
    BumpWindowDto:
      type: object
      properties:
        openedAt:
          $ref: "#/components/schemas/Instant"
        closesAt:
          $ref: "#/components/schemas/Instant"
        open:
          type: boolean
        outcome:
          type: string
        summary:
          type: string
        inFlight:
          type: integer
          format: int32
        slots:
          type: integer
          format: int32
        free:
          type: integer
          format: int32
        ciActive:
          type: integer
          format: int32
        owed:
          type: integer
          format: int32
        held:
          type: integer
          format: int32
        stalled:
          type: array
          items:
            $ref: "#/components/schemas/StalledBumpDto"
        queue:
          type: array
          items:
            $ref: "#/components/schemas/OwedBumpDto"
        next:
          type: string
        picks:
          type: array
          items:
            type: string
    Change:
      type: object
      properties:
        ecosystem:
          type: string
        manifestPath:
          type: string
        name:
          type: string
        from:
          type: string
        to:
          type: string
        location:
          type: string
...
    FailureDto:
      type: object
      properties:
        stepIndex:
          type: integer
          format: int32
        image:
          type: string
        exitCode:
          type: integer
          format: int32
        excerpt:
          type: string
...
    IngestRequest:
      type: object
      properties:
        ecosystem:
          type: string
        name:
          type: string
        version:
          type: string
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    OwedBumpDto:
      type: object
      properties:
        repository:
          type: string
        group:
          type: string
        changes:
          type: integer
          format: int32
        reason:
          type: string
        detail:
          type: string
    PackageDto:
      type: object
      properties:
        ecosystem:
          type: string
        name:
          type: string
    PendingBumpsDto:
      type: object
      properties:
        bumps:
          type: array
          items:
            $ref: "#/components/schemas/BumpDto"
...
    StalledBumpDto:
      type: object
      properties:
        repository:
          type: string
        group:
          type: string
        releaseRequestId:
          type: string
        state:
          type: string
        reason:
          type: string
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
  /maintenance/api/adoption/by-release:
    get:
      summary: How far one released repository and version got
      operationId: getAdoptionByRelease
      parameters:
      - name: repository
        in: query
        schema:
          type: string
      - name: version
        in: query
        schema:
          type: string
      responses:
        "200":
          description: The journey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdoptionJourneyDto"
        "400":
          description: Both repository and version are required
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Adoption Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/artifacts:
    get:
      summary: "Every artifact this platform publishes, with its reach"
      operationId: listArtifacts
      responses:
        "200":
          description: The artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtifactDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Artifact Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/artifacts/ingest:
    post:
      summary: Read one released artifact's sbom now
      operationId: ingestArtifactSbom
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequest"
        required: true
      responses:
        "202":
          description: Queued; the id is the artifact row
        "400":
          description: "Unknown ecosystem, or no name or version"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Artifact Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
  /maintenance/api/bumps:
    get:
      summary: "The newest bumps, of one repository or of all of them"
      operationId: listBumps
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      - name: repository
        in: query
        schema:
          type: string
      responses:
        "200":
          description: The bumps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BumpDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Bump Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/bumps/pending:
    get:
      summary: The newest bumps still on their way
      operationId: listPendingBumps
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: The pending bumps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingBumpsDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Bump Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/bumps/window:
    get:
      summary: The bump dispatch window and everything owed a bump
      operationId: getBumpWindow
      responses:
        "200":
          description: "The window, or the reason there is none"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BumpWindowDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Bump Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
    post:
      summary: Open a bump dispatch window now
      operationId: openBumpWindow
      responses:
        "200":
          description: The window that is now open
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BumpWindowDto"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Bump Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
    delete:
      summary: Close the bump dispatch window
      operationId: closeBumpWindow
      responses:
        "204":
          description: There is no window now
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Bump Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/bumps/{id}:
    get:
      summary: One bump with the changes it sent
      operationId: getBump
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The bump
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BumpDto"
        "404":
          description: No such bump
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Bump Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
...
info:
  title: qits-platform-maintenance API
  version: 2026.1002.201448